RunResult: Move to separate file and simplify
