+++ pisek/jobs/parts/run_result.py
@@ -1,0 +1,46 @@
+# pisek  - Tool for developing tasks for programming competitions.
+#
+# Copyright (c)   2023        Daniel Skýpala <daniel@honza.info>
+
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# any later version.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+from enum import Enum
+from typing import Optional, Union, Callable
+
+from pisek.paths import TaskPath
+
+class RunResultKind(Enum):
+    OK = 0
+    RUNTIME_ERROR = 1
+    TIMEOUT = 2
+
+class RunResult:
+    """Represents the way the program execution ended. Specially, a program
+    that finished successfully, but got Wrong Answer, still gets the OK
+    RunResult."""
+
+    def __init__(
+        self,
+        kind: RunResultKind,
+        returncode: int,
+        time: float,
+        wall_time: float,
+        stdout_file: Optional[Union[TaskPath, int]] = None,
+        stderr_file: Optional[TaskPath] = None,
+        stderr_text: Optional[str] = None,
+        status: str = "",
+    ):
+        self.kind = kind
+        self.returncode = returncode
+        self.stdout_file = stdout_file
+        self.stderr_file = stderr_file
+        self.stderr_text = stderr_text
+        self.status = status
+        self.time = time
+        self.wall_time = wall_time
+
